Move Legacy Widget block to @wordpress/widgets
