+++ docs/how-to-guides/widgets/legacy-widget-block.md
@@ -123,3 +123,49 @@ function hide_example_widget( $widget_types ) {
 }
 add_filter( 'widget_types_to_hide_from_legacy_widget_block', 'hide_example_widget' );
 ```
+
+## Using the Legacy Widget block in other block editors (Advanced)
+
+You may optionally allow the Legacy Widget block in other block editors such as
+the WordPress post editor. This is not enabled by default.
+
+First, ensure that any styles and scripts required by the legacy widgets are
+loaded onto the page. A convenient way of doing this is to manually perform all
+of the hooks that ordinarily run when a user browses to the widgets WP Admin
+screen.
+
+```php
+add_action( 'admin_print_styles', function() {
+	if ( get_current_screen()->is_block_editor() ) {
+		do_action( 'admin_print_styles-widgets.php' );
+	}
+} );
+add_action( 'admin_print_scripts', function() {
+	if ( get_current_screen()->is_block_editor() ) {
+		do_action( 'load-widgets.php' );
+		do_action( 'widgets.php' );
+		do_action( 'sidebar_admin_setup' );
+		do_action( 'admin_print_scripts-widgets.php' );
+	}
+} );
+add_action( 'admin_print_footer_scripts', function() {
+	if ( get_current_screen()->is_block_editor() ) {
+		do_action( 'admin_print_footer_scripts-widgets.php' );
+	}
+} );
+add_action( 'admin_footer', function() {
+	if ( get_current_screen()->is_block_editor() ) {
+		do_action( 'admin_footer-widgets.php' );
+	}
+} );
+```
+
+Then, register the Legacy Widget block using `registerLegacyWidgetBlock` which
+is defined in the `@wordpress/widgets` package.
+
+```php
+add_action( 'enqueue_block_editor_assets', function() {
+	wp_enqueue_script( 'wp-widgets' );
+	wp_add_inline_script( 'wp-widgets', 'wp.widgets.registerLegacyWidgetBlock()' );
+} );
+```
