+++ lib/widgets.php
@@ -11,6 +11,12 @@
  * @return boolean True if a screen containing the block editor is being loaded.
  */
 function gutenberg_is_block_editor() {
+	_deprecated_function(
+		'gutenberg_is_block_editor',
+		'10.8',
+		'WP_Screen::is_block_editor'
+	);
+
 	// If get_current_screen does not exist, we are neither in the standard block editor for posts, or the widget block editor.
 	// We can safely return false.
 	if ( ! function_exists( 'get_current_screen' ) ) {
@@ -45,79 +51,6 @@ function gutenberg_use_widgets_block_editor() {
 }
 
 /**
- * Emulates the Widgets screen `admin_print_styles` when at the block editor
- * screen.
- */
-function gutenberg_block_editor_admin_print_styles() {
-	if ( gutenberg_is_block_editor() ) {
-		/** This action is documented in wp-admin/admin-footer.php */
-		// phpcs:ignore WordPress.NamingConventions.ValidHookName.UseUnderscores
-		do_action( 'admin_print_styles-widgets.php' );
-	}
-}
-add_action( 'admin_print_styles', 'gutenberg_block_editor_admin_print_styles' );
-
-/**
- * Emulates the Widgets screen `admin_print_scripts` when at the block editor
- * screen.
- */
-function gutenberg_block_editor_admin_print_scripts() {
-	if ( gutenberg_is_block_editor() ) {
-		/** This action is documented in wp-admin/includes/ajax-actions.php */
-		do_action( 'load-widgets.php' ); // phpcs:ignore WordPress.NamingConventions.ValidHookName.UseUnderscores
-		/** This action is documented in wp-admin/includes/ajax-actions.php */
-		do_action( 'widgets.php' ); // phpcs:ignore WordPress.NamingConventions.ValidHookName.UseUnderscores
-		/** This action is documented in wp-admin/widgets.php */
-		do_action( 'sidebar_admin_setup' );
-		// phpcs:ignore WordPress.NamingConventions.ValidHookName.UseUnderscores
-		do_action( 'admin_print_scripts-widgets.php' );
-	}
-}
-add_action( 'admin_print_scripts', 'gutenberg_block_editor_admin_print_scripts' );
-
-/**
- * Emulates the Widgets screen `admin_print_footer_scripts` when at the block
- * editor screen.
- */
-function gutenberg_block_editor_admin_print_footer_scripts() {
-	if ( gutenberg_is_block_editor() ) {
-		/** This action is documented in wp-admin/admin-footer.php */
-		// phpcs:ignore WordPress.NamingConventions.ValidHookName.UseUnderscores
-		do_action( 'admin_print_footer_scripts-widgets.php' );
-	}
-}
-add_action( 'admin_print_footer_scripts', 'gutenberg_block_editor_admin_print_footer_scripts' );
-
-/**
- * Emulates the Widgets screen `admin_footer` when at the block editor screen.
- */
-function gutenberg_block_editor_admin_footer() {
-	if ( gutenberg_is_block_editor() ) {
-		/** This action is documented in wp-admin/admin-footer.php */
-		// phpcs:ignore WordPress.NamingConventions.ValidHookName.UseUnderscores
-		do_action( 'admin_footer-widgets.php' );
-	}
-}
-add_action( 'admin_footer', 'gutenberg_block_editor_admin_footer' );
-
-/**
- * Adds a save widgets nonce required by the legacy widgets block.
- */
-function gutenberg_print_save_widgets_nonce() {
-	// The function wpWidgets.save needs this nonce to work as expected.
-	echo implode(
-		"\n",
-		array(
-			'<form method="post">',
-			wp_nonce_field( 'save-sidebar-widgets', '_wpnonce_widgets', false ),
-			'</form>',
-		)
-	);
-}
-add_action( 'admin_footer-widgets.php', 'gutenberg_print_save_widgets_nonce' );
-
-
-/**
  * Returns the settings required by legacy widgets blocks.
  *
  * @return array Legacy widget settings.
@@ -137,7 +70,6 @@ function gutenberg_get_legacy_widget_settings() {
 			'media_image',
 			'media_gallery',
 			'media_video',
-			'meta',
 			'search',
 			'text',
 			'categories',
@@ -150,88 +82,10 @@ function gutenberg_get_legacy_widget_settings() {
 		)
 	);
 
-	// Backwards compatibility. Remove this in or after Gutenberg 10.5.
-	if ( has_filter( 'widgets_to_exclude_from_legacy_widget_block' ) ) {
-		/**
-		 * Filters the list of widget classes that should **not** be offered by the legacy widget block.
-		 *
-		 * Returning an empty array will make all the widgets available.
-		 *
-		 * @param array $widgets An array of excluded widgets classnames.
-		 *
-		 * @since 5.6.0
-		 */
-		$widgets_to_exclude_from_legacy_widget_block = apply_filters(
-			'widgets_to_exclude_from_legacy_widget_block',
-			array(
-				'WP_Widget_Block',
-				'WP_Widget_Pages',
-				'WP_Widget_Calendar',
-				'WP_Widget_Archives',
-				'WP_Widget_Media_Audio',
-				'WP_Widget_Media_Image',
-				'WP_Widget_Media_Gallery',
-				'WP_Widget_Media_Video',
-				'WP_Widget_Meta',
-				'WP_Widget_Search',
-				'WP_Widget_Text',
-				'WP_Widget_Categories',
-				'WP_Widget_Recent_Posts',
-				'WP_Widget_Recent_Comments',
-				'WP_Widget_RSS',
-				'WP_Widget_Tag_Cloud',
-				'WP_Widget_Custom_HTML',
-			)
-		);
-
-		_deprecated_hook(
-			'widgets_to_exclude_from_legacy_widget_block',
-			'10.3',
-			"wp.hooks.addFilter( 'legacyWidget.isWidgetTypeHidden', ... )"
-		);
-
-		foreach ( $wp_widget_factory->widgets as $widget ) {
-			if (
-				in_array( get_class( $widget ), $widgets_to_exclude_from_legacy_widget_block, true ) &&
-				! in_array( $widget->id_base, $widget_types_to_hide_from_legacy_widget_block, true )
-			) {
-				$widget_types_to_hide_from_legacy_widget_block[] = $widget->id_base;
-			}
-		}
-	}
-
 	$settings['widgetTypesToHideFromLegacyWidgetBlock'] = $widget_types_to_hide_from_legacy_widget_block;
 
 	return $settings;
 }
-
-/**
- * Extends default editor settings with values supporting legacy widgets.
- *
- * This can be removed when plugin support requires WordPress 5.8.0+.
- *
- * @param array $settings Default editor settings.
- *
- * @return array Filtered editor settings.
- */
-function gutenberg_legacy_widget_settings( $settings ) {
-	return array_merge( $settings, gutenberg_get_legacy_widget_settings() );
-}
-// This can be removed when plugin support requires WordPress 5.8.0+.
-if ( function_exists( 'get_block_editor_settings' ) ) {
-	add_filter( 'block_editor_settings_all', 'gutenberg_legacy_widget_settings' );
-} else {
-	add_filter( 'block_editor_settings', 'gutenberg_legacy_widget_settings' );
-}
-
-/**
- * Function to enqueue admin-widgets as part of the block editor assets.
- */
-function gutenberg_enqueue_widget_scripts() {
-	wp_enqueue_script( 'admin-widgets' );
-}
-
-add_action( 'enqueue_block_editor_assets', 'gutenberg_enqueue_widget_scripts' );
 
 /**
  * Overrides dynamic_sidebar_params to make sure Blocks are not wrapped in <form> tag.
@@ -306,4 +160,7 @@ function gutenberg_set_show_instance_in_rest_on_core_widgets() {
 		}
 	}
 }
-add_action( 'widgets_init', 'gutenberg_set_show_instance_in_rest_on_core_widgets' );
+
+if ( ! function_exists( 'wp_use_widgets_block_editor' ) ) {
+	add_action( 'widgets_init', 'gutenberg_set_show_instance_in_rest_on_core_widgets' );
+}
