+++ packages/widgets/src/blocks/legacy-widget/block.json
