+++ packages/widgets/src/blocks/legacy-widget/edit/control.js
