+++ packages/widgets/src/blocks/legacy-widget/edit/widget-type-selector.js
