+++ packages/widgets/src/blocks/legacy-widget/editor.scss
