+++ packages/widgets/src/blocks/legacy-widget/index.php
